Make unevenMix() ignore placeholders.
This is important in places like TypingLabelTest, which uses multiple non-color words in a color name, and those shouldn't change the resulting color. This also uses Java 8 method references to replace the GdxColorLookup class with a small method in ColorUtils that gets referenced in ColorLookup. GdxColorLookup wasn't used here anyway, and its features are all still here, just not the duplicated API of ColorLookup.INSTANCE and ColorLookup.GdxColorLookup.INSTANCE.