Lots of clearing up in the README.md . Regex fix.
The monstrous COLOR_MARKUP_TO_TAG regex previously tolerated any character a Java identifier can start with for the first letter in a color name. That would potentially conflict with `[_]` for underline, so now it tolerates any alphanumeric Unicode character in the first position, but nothing else in that position.