Fix color parsing with spaces.
This is a weird one, but it doesn't change the fix in the last commit. The check for whether a color name was used didn't have any way to check if adjectives or quantifiers were present in a color description String, so it just ignored any colors that started with `light`, `dull`, etc. It probably ignored quite a lot of other things, too. Now it works!