Important fix for state storage in TypingLabel.
Because TypingLabel converts between square bracket and curly brace markup a few times, it's vital that all square bracket markup can be brought back after conversion to curly brackets. This is possible now, but wasn't for a while. The `[(storage)]` and `[ storage]` save and load syntaxes were ignored when converting to curly brackets, and now this is fixed.