Test out ArrayTextureSpriteBatch.
This is related to https://github.com/libgdx/libgdx/pull/5914 . I've found that the framerate is significantly higher with ArrayTextureSpriteBatch when compared to SpriteBatch, but both are lower than just using SpriteBatch on GL20. GL20 SpriteBatch is almost twice as fast as GL30 ArrayTextureSpriteBatch, which seems to suggest something's not right in GL30 or my limited usage of it.