Attempt to test a different Batch.
This is Nathan Sweet's set of edits to an older version of ArrayTextureSpriteBatch, from https://n4te.com/x/9884-ICFL.txt . That page is unfortunately **nigh-unusable** if you're trying to copy code out of it. That is why the formatting on comments is completely wrecked here, even after extensive manual editing... Why Nate wouldn't just paste an actual .txt file is beyond me.

As for the actual content of the new Batch, it isn't really much different from SpriteBatch here. It might be a little faster or a little slower than SpriteBatch. I do have to call getters because so much of the libGDX graphics code is package-private, so if that's an issue here, it wouldn't be in libGDX core.

God, I hate package-private variables.