Get rid of some hacky stuff.
In resizeDistanceField(), there were some ad-hoc adjustments to MSDF and SDF crispness, and they were never really quite right. They're much simpler now. In enableShader() and enableDistanceFieldShader(), it was failing to set uniforms if the u_smoothing value was the same as the value used for the last thing, even if the shader was different (and so the stored value couldn't be the same).