Figured out a maddening issue with emoji.
Positions were different depending on if a Font had had emoji added before or after scaling to the desired size. This was not the ideal behavior, and it stemmed from usage of descent * scaleY in the offsetY in calls to addAtlas(). Because scaleY changes when the font is resized, that made the y-position depend on when emoji (or other atlases) were added.